refactor: Separate the 'SLevelController' controller between multiple files.
Now being a partial static class, its code was divided into several files arranged by names that characterize that region of code and its components.
